+++ Editor/MenuItemLayout.cs
@@ -7,10 +7,16 @@ namespace Baracuda.Monitoring.Editor
 {
     internal class MenuItemLayout : MonoBehaviour
     {
-        [MenuItem("Tools/Runtime Monitoring", priority = 2405)]
+        [MenuItem("Tools/Runtime Monitoring/Settings", priority = 2405)]
         private static void OpenMonitorSettingsWindow()
         {
             MonitoringSettingsWindow.Open();
+        }
+
+        [MenuItem("Tools/Runtime Monitoring/Filter Window", priority = 2406)]
+        private static void OpenFiltersWindow()
+        {
+            MonitoringFilterWindow.Open();
         }
     }
 }